beta

+++ src/main/java/com/spring/security/service/impl/S3UploadServiceImpl.java
@@ -6,22 +6,22 @@ import java.io.IOException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.web.multipart.MultipartFile;
 import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
+import org.springframework.web.multipart.MultipartFile;
 
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.spring.security.service.S3UploadService;
 
+import lombok.RequiredArgsConstructor;
+
 @Service
 @RequiredArgsConstructor
 public class S3UploadServiceImpl implements S3UploadService {
 
     @Autowired
-    private AmazonS3 s3Client;
+    private final AmazonS3 s3Client;
 
     @Value("${application.bucket.name}")
     private String bucketName;
@@ -31,53 +31,62 @@ public class S3UploadServiceImpl implements S3UploadService {
 
     @Value("${cloud.aws.s3.endpoint}")
     private String endPoint;
-    
-    String fileName;
 
+    String fileName;
 
     @Override
     public String uploadImage(final MultipartFile multiplePartToFile, final String folderName) {
-        try{
-            final File file= convertMultiPartFileToFile(multiplePartToFile);
-            
+        try {
+            final File file = convertMultiPartFileToFile(multiplePartToFile);
+            System.out.println("Uploading to bucket: " + bucketName);
+            System.out.println("Region: " + region);
+            System.out.println("File name: " + file.getName());
+
+            if (!doesBucketExist(bucketName)) {
+                throw new RuntimeException("Bucket " + bucketName + " does not exist.");
+            }
+
             fileName = uploadFileToS3Bucket(bucketName, folderName, file);
+            System.out.println("Uploaded file name: " + fileName);
             file.delete();
 
-       
-        }catch(final AmazonServiceException e){
-            System.out.println("Error Message: " + e.getErrorMessage());
+        } catch (final AmazonServiceException e) {
+            System.out.println("Amazon Service Error: " + e.getErrorMessage());
+        } catch (final Exception e) {
+            System.out.println("General Error: " + e.getMessage());
         }
         return fileName;
     }
 
-
     @Override
     public String deleteImage(final String fileName) {
-       try{
-         s3Client.deleteObject(bucketName, fileName);
-        return fileName;
-       }catch(final AmazonServiceException e){
-           System.out.println("Error Message:    " + e.getErrorMessage());
-           return null;
-       }
+        try {
+            s3Client.deleteObject(bucketName, fileName);
+            return fileName;
+        } catch (final AmazonServiceException e) {
+            System.out.println("Error Message: " + e.getErrorMessage());
+            return null;
+        }
     }
 
     private File convertMultiPartFileToFile(final MultipartFile multiplePartToFile) {
-       File file = new File(multiplePartToFile.getOriginalFilename());
-       try(final FileOutputStream fos = new FileOutputStream(file)){
-           fos.write(multiplePartToFile.getBytes());
-       }catch(final IOException e){
-           System.out.println("Error Message: " + e.getMessage());
-       }
-       return file;
+        File file = new File(multiplePartToFile.getOriginalFilename());
+        try (final FileOutputStream fos = new FileOutputStream(file)) {
+            fos.write(multiplePartToFile.getBytes());
+        } catch (final IOException e) {
+            System.out.println("Error Message: " + e.getMessage());
+        }
+        return file;
     }
 
-    private String uploadFileToS3Bucket(final String bucketName,final String folder, final File file) {
-        fileName =  folder + "/" + System.currentTimeMillis() + "_" + file.getName().replaceAll("\\s+", "_");
-        final PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, fileName, file);   
-
+    private String uploadFileToS3Bucket(final String bucketName, final String folder, final File file) {
+        fileName = folder + "/" + System.currentTimeMillis() + "_" + file.getName().replaceAll("\\s+", "_");
+        final PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, fileName, file);
         s3Client.putObject(putObjectRequest);
         return fileName;
     }
-    
+
+    private boolean doesBucketExist(String bucketName) {
+        return s3Client.doesBucketExistV2(bucketName);
+    }
 }
